moved Schema to core
